requirements.txt
# =============================================================================
# EPUB2PDF - Dépendances principales
# =============================================================================

# Interface utilisateur moderne (requis)
PySide6>=6.5.0

# Traitement d'images et conversion PDF (requis)
Pillow>=10.0.0
PyPDF2>=3.0.0

# Extraction d'archives (requis)
rarfile>=4.0

# Requêtes HTTP pour métadonnées (requis)
requests>=2.28.0

# =============================================================================
# Dépendances optionnelles pour de meilleures performances
# =============================================================================

# Traitement d'images avancé (optionnel - améliore la qualité)
wand>=0.6.0

# Optimisation des performances (optionnel)
numba>=0.58.0
psutil>=5.9.0

# =============================================================================
# Dépendances de développement et tests
# =============================================================================

# Tests
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-qt>=4.2.0

# Qualité de code et formatage
coverage>=7.0.0
black>=22.0.0
flake8>=5.0.0

# =============================================================================
# Notes d'installation
# =============================================================================
# 
# Installation complète (avec dépendances optionnelles) :
# pip install -r requirements.txt
#
# Installation minimale (dépendances requises uniquement) :
# pip install PySide6>=6.5.0 Pillow>=10.0.0 PyPDF2>=3.0.0 rarfile>=4.0 requests>=2.28.0
#
# Dépendances système requises :
# - unar (pour l'extraction d'archives) : brew install unar (macOS) ou apt-get install unar (Ubuntu)
# - ImageMagick (pour wand) : brew install imagemagick (macOS) ou apt-get install imagemagick (Ubuntu) 